requirements.txt
aiohttp==3.8.5
beautifulsoup4==4.11.1
Flask==3.1.0
Flask_Cors==5.0.0
instaloader==4.14
numpy==1.25.2
opencv_python_headless==4.10.0.84
pandas==2.2.3
requests==2.32.3
catboost==1.2.7
scikit-learn==1.5.2
google-reverse-search==0.1.3
google-image-source-search==1.2.2
googlesearch-python==1.2.5